fix: divided routing and markup
